feat(bot,ci): rename telegram_bot to bot, delete teams_analyzer
PR 5/6 of the biwenger-api refactor. The big destructive step:
* The Telegram bot package is now packages/biwenger_tools/bot/.
* The teams_analyzer package and Cloud Run Job stop existing
  (everything was already moved to biwenger-api in PR 2–PR 3).

Code rename:
* packages/biwenger_tools/telegram_bot → packages/biwenger_tools/bot
  via git mv (history preserved).
* Import paths updated: packages.biwenger_tools.telegram_bot → .bot.
* Bazel target name: telegram_bot → bot; image repository →
  europe-southwest1-docker.pkg.dev/.../biwenger-docker/bot.
* Tests file renamed test_telegram_bot.py → test_bot.py.

Deletions:
* packages/biwenger_tools/teams_analyzer/ — entire package.

CI (deploy.yml):
* "teams_analyzer" path filter and detect-changes output removed.
* "telegram_bot" filter renamed to "bot", path updated.
* deploy-teams-analyzer job removed.
* deploy-telegram-bot job renamed to deploy-bot. Cloud Run service
  name: biwenger-telegram-bot → biwenger-bot. Image path: /telegram_bot
  → /bot.
* test job: removed teams_analyzer_tests, renamed telegram_bot_tests
  → bot_tests.
* cleanup needs: deploy-teams-analyzer removed, deploy-telegram-bot
  → deploy-bot.

scripts/clean-images-artifact.sh:
* SIMPLE_IMAGES: removed "teams_analyzer" and "telegram_bot", added
  "bot" (api was already added in PR 1).

Manual steps required after this PR merges (same session):
* Update Telegram webhook URL to point to biwenger-bot's new URL.
* Delete the old biwenger-telegram-bot Cloud Run service.
* Delete the biwenger-teams-analyzer Cloud Run Job.
* (Cosmetic) rename the bot's display name to "Biwenger Bot" in
  BotFather.

PR 6 will sweep READMEs and docs that still mention teams_analyzer
or telegram_bot.
