+++ packages/biwenger_tools/bot/gunicorn_prod_runner.py
@@ -1,4 +1,4 @@
-"""Gunicorn launcher for the telegram bot service on Cloud Run."""
+"""Gunicorn launcher for the Biwenger bot service on Cloud Run."""
 
 import sys
 
@@ -9,6 +9,6 @@ if __name__ == "__main__":
         "gunicorn",
         "--bind",
         "0.0.0.0:8080",
-        "packages.biwenger_tools.telegram_bot.app:app",
+        "packages.biwenger_tools.bot.app:app",
     ]
     run()
