+++ packages/biwenger_tools/bot/requirements.txt
@@ -1,4 +1,4 @@
-# telegram_bot/requirements.txt
+# bot/requirements.txt
 # All deps (flask, gunicorn, python-dotenv, google-auth, requests) are already
 # present in requirements_lock.txt via core/ and web/. No new additions needed.
 Flask
