--- packages/biwenger_tools/teams_analyzer/BUILD.bazel
@@ -1,152 +1,0 @@
-# teams_analyzer/BUILD.bazel
-load("@rules_pkg//pkg:tar.bzl", "pkg_tar")
-load("@rules_pkg//pkg:mappings.bzl", "strip_prefix")
-load("@rules_python//python:defs.bzl", "py_binary", "py_library", "py_test")
-load("@rules_oci//oci:defs.bzl", "oci_image", "oci_load", "oci_push")
-
-# =================================================================
-# 1️⃣ MAIN LIBRARY
-# =================================================================
-py_library(
-    name = "teams_analyzer_lib",
-    srcs = glob(
-        ["**/*.py"],
-        exclude = ["tests/**/*.py"],
-    ),
-    deps = [
-        "//core",
-        "@pypi//matplotlib",
-        "@pypi//python_dotenv",
-        "@pypi//requests",
-        "@pypi//unidecode",
-    ],
-)
-
-# =================================================================
-# 2️⃣ LOCAL EXECUTION (NO DOCKER)
-# =================================================================
-py_binary(
-    name = "teams_analyzer_local",
-    main = "main.py",
-    srcs = ["main.py"],
-    data = [".env"],
-    deps = [":teams_analyzer_lib"],
-)
-
-# =================================================================
-# 3️⃣ DOCKER IMAGE LAYERS
-# =================================================================
-
-# Core: pre-tarred sources to preserve directory structure
-pkg_tar(
-    name = "core_layer",
-    files = {
-        "//core:core_srcs": "core_srcs.tar",
-    },
-    package_dir = "/app",
-)
-
-# Root .py files for the package
-pkg_tar(
-    name = "_code_root",
-    srcs = glob(["*.py"]) + ["entrypoint.sh"],
-    package_dir = "/app/packages/biwenger_tools/teams_analyzer",
-    strip_prefix = strip_prefix.from_pkg(),
-    mode = "0755",
-)
-
-# Subdirectory logic/
-pkg_tar(
-    name = "_code_logic",
-    srcs = glob(["logic/*.py"]),
-    package_dir = "/app/packages/biwenger_tools/teams_analyzer/logic",
-    strip_prefix = strip_prefix.from_pkg("logic"),
-    mode = "0755",
-)
-
-# Local-only: include the .env so the binary can read credentials
-pkg_tar(
-    name = "_secrets",
-    srcs = [".env"],
-    package_dir = "/app/packages/biwenger_tools/teams_analyzer",
-    strip_prefix = strip_prefix.from_pkg(),
-)
-
-# GCP-bound layer (no secrets — Cloud Run injects them at runtime)
-pkg_tar(
-    name = "teams_analyzer_base_layer",
-    deps = [":_code_root", ":_code_logic"],
-)
-
-# Local layer (with secrets baked in — only for local docker run)
-pkg_tar(
-    name = "teams_analyzer_local_layer",
-    deps = [":_code_root", ":_code_logic", ":_secrets"],
-)
-
-# =================================================================
-# 4️⃣ DOCKER IMAGES (LOCAL AND GCP)
-# =================================================================
-
-# Local image with secrets, for `docker run` validation
-oci_image(
-    name = "teams_analyzer_image_local",
-    base = "@python_with_deps",
-    tars = [
-        ":core_layer",
-        ":teams_analyzer_local_layer",
-    ],
-    env = {
-        "PYTHONPATH": "/app",
-        "PYTHONUNBUFFERED": "1",
-    },
-    entrypoint = ["/app/packages/biwenger_tools/teams_analyzer/entrypoint.sh"],
-)
-
-# GCP image without secrets — Cloud Run / Cloud Run Job injects them
-oci_image(
-    name = "teams_analyzer_image_gcp",
-    base = "@python_with_deps",
-    tars = [
-        ":core_layer",
-        ":teams_analyzer_base_layer",
-    ],
-    env = {
-        "PYTHONPATH": "/app",
-        "PYTHONUNBUFFERED": "1",
-    },
-    entrypoint = ["/app/packages/biwenger_tools/teams_analyzer/entrypoint.sh"],
-)
-
-# =================================================================
-# 5️⃣ DOCKER + GCP COMMANDS
-# =================================================================
-
-oci_load(
-    name = "load_image_to_docker_local",
-    image = ":teams_analyzer_image_local",
-    repo_tags = ["bazel/teams_analyzer:local"],
-)
-
-oci_push(
-    name = "push_image_to_gcp",
-    image = ":teams_analyzer_image_gcp",
-    repository = "europe-southwest1-docker.pkg.dev/biwenger-tools/biwenger-docker/teams_analyzer",
-    remote_tags = ["latest"],
-)
-
-# =================================================================
-# 6️⃣ UNIT TESTS (PYTEST)
-# =================================================================
-py_test(
-    name = "teams_analyzer_tests",
-    args = ["packages/biwenger_tools/teams_analyzer/tests/"],
-    timeout = "short",
-    main = "tests/main.py",
-    srcs = glob(["tests/**/*.py"]),
-    deps = [
-        ":teams_analyzer_lib",
-        "//core",
-        "@pypi//pytest",
-    ],
-)
