--- packages/biwenger_tools/teams_analyzer/entrypoint.sh
@@ -1,16 +1,0 @@
-#!/bin/sh
-set -e
-
-# Extract the 'core' package if present in the root directory.
-# Necessary because Bazel packages it as a .tar to preserve directory structure.
-if [ -f /app/core_srcs.tar ]; then
-    echo ">>> Initializing: Extracting core package..."
-    cd /app
-    tar -xf core_srcs.tar
-    rm -f core_srcs.tar
-fi
-
-export PYTHONPATH="/app:${PYTHONPATH}"
-
-echo ">>> Starting teams analyzer..."
-exec python3 -m packages.biwenger_tools.teams_analyzer.main
