--- packages/biwenger_tools/teams_analyzer/requirements.txt
@@ -1,5 +1,0 @@
-# teams_analyzer/requirements.txt
-python-dotenv
-requests
-unidecode
-matplotlib
