--- packages/biwenger_tools/teams_analyzer/tests/main.py
@@ -1,6 +1,0 @@
-import sys
-import pytest
-
-# Lanza pytest con los argumentos del sistema y sale con el código de estado correcto
-if __name__ == "__main__":
-    sys.exit(pytest.main())
